requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
openpyxl
python-dotenv
